requirements.txt
python-telegram-bot==13.15
pyopenssl==22.1.0
urllib3==1.26.17
python-dotenv
pytimeparse